requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary
alembic

# HTTP Client
httpx

# Authentication
pyjwt
passlib[bcrypt]
bcrypt==4.0.1

# Validation
pydantic
email-validator

# Environment
python-dotenv

# MCP SDK
mcp
